requirements/base.txt
# base.txt is a pip requirements file which describes the necessary
# distributions required to run this project. If you need something that is
# project specific not listed here use project.txt. You can, of course, update
# versions of distributions here if needed.

--extra-index-url=http://dist.pinaxproject.com/dev/
--extra-index-url=http://dist.pinaxproject.com/alpha/
--extra-index-url=http://dist.pinaxproject.com/fresh-start/

Django==1.3.7
Pinax

django-debug-toolbar==0.8.5
django-staticfiles==1.1.2
django_compressor==1.0.1

django-mailer==0.2a1
django-email-confirmation==0.2
django-timezones==0.2
pytz==2011n
django-openid==0.3a1
python-openid==2.2.5

django-announcements==0.1.2
django-pagination==1.0.7
django-notification==0.2
# =todo: upgrade to django 1.4
# git+https://github.com/hypertexthero/django-notification#egg=django-notification
https://github.com/yourcelf/django-notification
idios==0.2a1
metron==0.1

# pinax-theme-bootstrap==0.1.2
